instant-dashboard/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
groq==0.11.0
langchain==0.3.13
langchain-groq==0.2.1
langchain-core==0.3.28
pydantic==2.9.2
python-dotenv==1.0.1
python-multipart==0.0.12
